Move `WithComputeUnitPrice` to `compute_budget`
