+++ cli/src/lib.rs
@@ -28,7 +28,6 @@ pub mod clap_app;
 pub mod cli;
 pub mod cluster_query;
 pub mod compute_budget;
-pub mod compute_unit_price;
 pub mod feature;
 pub mod inflation;
 pub mod memo;
